Split up CommandModal into stateless and stateful components.
