--- apps/dapp/components/CommandModal/index.tsx
@@ -1,4 +1,0 @@
-// GNU AFFERO GENERAL PUBLIC LICENSE Version 3. Copyright (C) 2022 DAO DAO Contributors.
-// See the "LICENSE" file in the root directory of this package for more copyright information.
-
-export * from './CommandModal'
