+++ packages/types/ui/index.ts
@@ -1,6 +1,7 @@
 export * from './ActionCardLoader'
 export * from './AppLayout'
 export * from './Breadcrumbs'
+export * from './CommandModal'
 export * from './common'
 export * from './DaoCard'
 export * from './DaoDropdown'
@@ -8,6 +9,7 @@ export * from './DaoInfoBar'
 export * from './HeroStat'
 export * from './Loader'
 export * from './Logo'
+export * from './Modal'
 export * from './Navigation'
 export * from './ProfileCardWrapper'
 export * from './ProfileNewProposalCard'
