+++ packages/ui/components/CommandModal/CommandHits.tsx
@@ -1,20 +1,11 @@
 // GNU AFFERO GENERAL PUBLIC LICENSE Version 3. Copyright (C) 2022 DAO DAO Contributors.
 // See the "LICENSE" file in the root directory of this package for more copyright information.
 
-import clsx from 'clsx'
-import { Fragment, forwardRef, useCallback, useEffect, useState } from 'react'
-import { useTranslation } from 'react-i18next'
+import { Fragment, useCallback, useEffect, useState } from 'react'
 
-import { Loader } from '@dao-dao/ui'
+import { Hit, HitSectionData } from '@dao-dao/tstypes/ui/CommandModal'
 
-import { DaoHit, Hit, HitType } from './CommandModal'
-
-interface HitSectionData {
-  // End index of each section, exclusive. For example, if the first section has
-  // 3 items in it, sectionEndIndexes[0] === 3.
-  sectionEndIndexes: number[]
-  sectionNames: string[]
-}
+import { HitRow } from './HitRow'
 
 export interface CommandHitsProps {
   sectionData: HitSectionData
@@ -93,7 +84,7 @@ export const CommandHits = ({
     <div className="flex overflow-y-auto flex-col grow gap-1 p-3 pt-4 no-scrollbar">
       {/* If hit we're currently navigating to is no longer part of the hits to render, just display at the top with a loader. */}
       {navigatingFromHit && !hits.includes(navigatingFromHit) && (
-        <HitView
+        <HitRow
           hit={navigatingFromHit}
           loading
           onClick={() => null}
@@ -114,10 +105,10 @@ export const CommandHits = ({
 
             {hits
               .slice(sectionStartIndex, sectionEndIndex)
-              .map((hit: DaoHit, hitIndex: number) => {
+              .map((hit: Hit, hitIndex: number) => {
                 const selected = sectionStartIndex + hitIndex === selection
                 return (
-                  <HitView
+                  <HitRow
                     key={hit.id}
                     hit={hit}
                     loading={navigatingFromHit === hit}
@@ -134,51 +125,3 @@ export const CommandHits = ({
     </div>
   )
 }
-
-interface HitViewProps {
-  hit: Hit
-  selected: boolean
-  onClick: () => void
-  loading: boolean
-}
-
-const HitView = forwardRef<HTMLDivElement, HitViewProps>(function HitView(
-  { hit, selected, onClick, loading },
-  ref
-) {
-  const { t } = useTranslation()
-
-  return (
-    <div
-      className={clsx(
-        'group flex flex-row gap-2 items-center p-3 bg-transparent hover:bg-background-interactive-hover rounded-md transition cursor-pointer',
-        selected && 'bg-background-interactive-hover'
-      )}
-      onClick={onClick}
-      ref={ref}
-    >
-      {hit.hitType === HitType.Dao ? (
-        <div
-          aria-label={t('info.daosLogo')}
-          className="w-[24px] h-[24px] bg-center bg-cover rounded-full"
-          role="img"
-          style={{
-            backgroundImage: `url(${hit.imageUrl})`,
-          }}
-        ></div>
-      ) : (
-        <div className="flex justify-center items-center w-[24px] h-[24px] text-lg">
-          {hit.icon}
-        </div>
-      )}
-
-      <p className="font-medium body-text">{hit.name}</p>
-
-      {loading && (
-        <div className="flex flex-row grow justify-end items-center">
-          <Loader fill={false} size={20} />
-        </div>
-      )}
-    </div>
-  )
-})
