+++ packages/ui/components/CommandModal/index.tsx
@@ -1,0 +1,1 @@
+export * from './CommandModal'
